Updated the Camera API.
Objects are now positioned/drawn where they should be. The
glfw part worries about where the camera is, and where the
objects should be rendered.

Added some tests to ensure its behavior doesn't change over time.

Removed the Clamp function in the math file, because it's
probably better to use the mgl32 package for things like that. No
need to define it twice.

Renamed the "gl" variable in engi, because having both gl and Gl,
referencing to the exact same object, doesn't make any sense.