Added KeyboardScroller
It's a system that can be called by the user, to listen for certain
keys, and then move in a certain direction with the given speed.

I also updated the KeyManager because of it; no need to hardcode
every key in three places in our code. I left some backwards
compatibility there.