Implemented PauseSystem and updated Message
[BREAKING CHANGE]
MessageManager.Dispatch now only requires the Message, no Name string.
This is because this information is already in Message.Type()

Handler now is no longer func(i interface{}), but func(msg Message),
since we know it'll be a message - it's best to use type safety.
It's also renamed to MessageHandler, to be more specific.

Fixed the examples so it didn't break anything.

Added demo for the PauseSystem.