Added MouseEvents AND changed window resize
Also added a demo for the mouse events. See #59.

The function engi.SetScaleOnResize allows user to say:
"I want things to become bigger on resize" (i.e. Pong), or
"I want to be able to see more things on resize" (i.e. KeyboardScroller).
See #65.

The MouseSystem and CameraSystem are changed accordingly.

Also created a separate audio.go file for Windows, to
be sure it also builds on Windows. Fixes #121.