Fixed some global variables wrt Scene
Fixes #126 and fixes #128.

Global variables are now updated on Scene creation (whenever
we initialize a new World instance). These values are saved
in a sceneWrapper, so we don't always have to reinitialize.

CameraSystem is now set on Scene creation, since this was
hard to do within RenderSystem (as it is Scene-specific).