Rotation support (#216)
Adds rotation support and texture cleanup

It also adds some features w.r.t. repeating textures, see issue #237.

Texture cleanup methods (.Close()) are added, which fixes #211 and fixes #135.

Shaders are reworked so they're easier to implement (also documented at #237). Buffers are now no longer handled by the RenderComponent. As a result, Scale is now a variable instead of a method, allowing everyone to access (and change) it directly.

Rotation is enabled for both RenderComponents and for the Camera. Also added demo to show the behavior and updated the Mouse demo to show it works with rotation as well.

A lot of credit goes to @faide for doing the groundwork for the rotation support.