font fixes
the diplayed font is now fully displayed. However, there is a slight pixel offset between letters with particularly bad overhangs. For example, e has a dot to the top right because f comes after it and the crossbar on the f gets drawn too. Possibly due to rounding. Showed up before these fixes and shows up on master branch.